langchain_voicebot/requirements.txt
openai=0.27.0
langchain==0.0.174
pyttsx3==2.90
python-dotenv==0.221.0